Refactor: Modularize core grammar components
Decomposed the large `JanusAI/core/grammar/base_grammar.py` file into smaller, more focused modules:
- `cfg.py` for CFGRule and ContextFreeGrammar
- `denoiser.py` for NoisyObservationProcessor
- `progressive_grammar.py` for ProgressiveGrammar
- `ai_grammar.py` for AIGrammar

Updated `base_grammar.py` to act as a facade, re-exporting these classes to maintain backward compatibility.

Adjusted imports throughout the project to point to the new specific module locations where appropriate. `BaseGrammar` imports were updated to use `ProgressiveGrammar` under the assumption it was an alias or old name.

Added/verified module and class docstrings for the new and moved components.

Note: Tests were not run prior to this commit due to an environmental disk space issue.
