+++ JanusAI/ai_interpretability/symbolic/neural_grammar.py
@@ -16,7 +16,7 @@
 
 import sympy as sp
 from typing import List, Dict, Any, Optional
-from JanusAI.core.grammar.base_grammar import ProgressiveGrammar
+from JanusAI.core.grammar.progressive_grammar import ProgressiveGrammar # Updated import
 from JanusAI.core.expressions.expression import Expression, Variable
 
 class AIGrammar(ProgressiveGrammar):
