+++ JanusAI/core/grammar/__init__.py
@@ -1,8 +1,19 @@
 # JanusAI/core/grammar/__init__.py
-from .base_grammar import BaseGrammar, ProgressiveGrammar, NoisyObservationProcessor
+from .progressive_grammar import ProgressiveGrammar
+from .ai_grammar import AIGrammar
+from .cfg import CFGRule, ContextFreeGrammar
+from .denoiser import NoisyObservationProcessor
+from .expression import Expression, Variable
 
+# Assuming BaseGrammar was an alias or precursor to ProgressiveGrammar.
+# If BaseGrammar was distinct and needed, this will need adjustment.
+# For now, ProgressiveGrammar is the main export.
 __all__ = [
-    "BaseGrammar",
     "ProgressiveGrammar",
+    "AIGrammar",
+    "CFGRule",
+    "ContextFreeGrammar",
     "NoisyObservationProcessor",
+    "Expression",
+    "Variable",
 ]
